requirements.txt
streamlit
deta
python-dotenv
discord-webhook